kafka: move member_protocol into server out of protocol
the vector overloads for equality are also removed since it appears that
only fragmented vectors are used now.
